+++ test/helloworld/src/hello.clj
@@ -1,5 +1,0 @@
-(ns hello)
-
-(defn -main [& _]
-  (println "Hello World!")
-  (println (slurp "/app/version.edn")))